feat(chat): redirect to base route on logo click
