+++ apps/chat/src/components/Header/Logo.tsx
@@ -1,0 +1,47 @@
+import { useRouter } from 'next/router';
+
+import classNames from 'classnames';
+
+import { ApiUtils } from '@/src/utils/server/api';
+
+import { useAppSelector } from '@/src/store/hooks';
+import { SettingsSelectors } from '@/src/store/settings/settings.reducers';
+import { UISelectors } from '@/src/store/ui/ui.reducers';
+
+import { Feature } from '@epam/ai-dial-shared';
+import cssEscape from 'css.escape';
+
+export const Logo = () => {
+  const router = useRouter();
+
+  const customLogo = useAppSelector(UISelectors.selectCustomLogo);
+  const isCustomLogoFeatureEnabled = useAppSelector((state) =>
+    SettingsSelectors.isFeatureEnabled(state, Feature.CustomLogo),
+  );
+
+  const customLogoUrl =
+    isCustomLogoFeatureEnabled &&
+    customLogo &&
+    `/api/${ApiUtils.encodeApiUrl(customLogo)}`;
+
+  const isClickable = router.route !== '/';
+
+  const handleLogoClick = () => {
+    if (isClickable) router.push('/');
+  };
+
+  return (
+    <span
+      onClick={handleLogoClick}
+      className={classNames(
+        'mx-auto min-w-[110px] bg-contain bg-center bg-no-repeat md:ml-5 lg:bg-left',
+        isClickable && 'cursor-pointer',
+      )}
+      style={{
+        backgroundImage: customLogoUrl
+          ? `url(${cssEscape(customLogoUrl)})`
+          : `var(--app-logo)`,
+      }}
+    ></span>
+  );
+};
